lambda/requirements.txt
Pillow==10.3.0
boto3==1.34.0
